% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfnct00.R
\name{cons}
\alias{cons}
\title{Constraint function}
\usage{
cons(yd, nu)
}
\arguments{
\item{yd}{Vector of derivatives.}

\item{nu}{Constraint parameter.}
}
\value{
Constraint value, as nu gets large returns 0 if yd is negative and 1 otherwise.
}
\description{
Constraint function
}
