% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfnct00.R
\name{llik}
\alias{llik}
\title{GP log-likelihood}
\usage{
llik(y_obs, R, sig2)
}
\arguments{
\item{y_obs}{Vector of outputs.}

\item{R}{Correlation matrix.}

\item{sig2}{scalar variance.}
}
\value{
log-likelihood
}
\description{
evaluates GP log-likelihood for a set of responses given variance and correlation matrix
}
