% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfnct00.R
\name{lmvnlik}
\alias{lmvnlik}
\title{Log Multivariate Gausssian likelihood}
\usage{
lmvnlik(y, mean, covmat)
}
\arguments{
\item{y}{Vector at which the pdf is obtained.}

\item{mean}{Vector mean of the multivariate Gaussian. Same length as \code{y}.}

\item{covmat}{covariance matrix of the multivariate Gaussian.}
}
\value{
Log multivariate Gaussian pdf with mean \code{mean} and covariance matrix \code{covmat}
for vector \code{y}
}
\description{
\code{lmvnlik} returns the log pdf of the multivariate normal. There are other functions that
do the exact same thing. But there are some manual error preventions here that might make
your life easier; or not!
}
