% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfnct00.R
\name{logprior}
\alias{logprior}
\title{Log Prior on GP parameters}
\usage{
logprior(l, sig2, lmean = 1, sigmean = 5)
}
\arguments{
\item{l}{Vector of length scale parameters.}

\item{sig2}{Variance. scalar.}

\item{lmean}{The mean of the chi-squared prior on \code{l}. Default value is 1.}

\item{sigmean}{The mean of the chi-squared prior on \code{l}. Default value is 5.}
}
\value{
Log-prior density at the given \code{l} and \code{sig2} pair.
}
\description{
Evaluates log-prior for a set of GP parameters
}
