% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfnct00.R
\name{lpart}
\alias{lpart}
\title{Update function}
\usage{
lpart(l, x, y_obs, newx)
}
\arguments{
\item{l}{Vector of length scale parameters.}

\item{x}{Design matrix.}

\item{newx}{Prediction points.}

\item{y}{Vector of responses.}
}
\value{
list of correlation matrix for the design points; mean vector and covariance
matrix of the posterior predictive
}
\description{
Updates the objects in sampling that are a function of the length scale \code{l}
}
