% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfnct00.R
\name{lpost}
\alias{lpost}
\title{Log posterior function}
\usage{
lpost(l, sig2, y, yd, corrmat, mean, R, nu, y_obs)
}
\arguments{
\item{l}{Correlation parameter}

\item{sig2}{Variance}

\item{y}{Vector of responses}

\item{yd}{Vector of derivatives}

\item{corrmat}{Posterior predictive correlation matrix}

\item{mean}{GP mean vector}

\item{R}{Likelihood correlation matrix}

\item{nu}{Constraint parameter}
}
\value{
log posterior predictive value
}
\description{
Log posterior function
}
