% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfnct00.R
\name{matcor}
\alias{matcor}
\title{Correlation matrices}
\usage{
matcor(x, y, l)
}
\arguments{
\item{x}{Matrix whose each row is a point in \code{R^D}}

\item{y}{Matrix whose each row is a point in \code{R^D}. The number of columns must be
the same as the number of columns in \code{x}.}

\item{l}{Vector of length scale parameters for each input dimension. Length of \code{l} must
be equal to the number of columns of \code{x} and \code{y}}
}
\value{
Correlation matrix between points in \code{x} and \code{y}.
}
\description{
\code{matcor} returns the correlation matrix between two sets of points given by the
matrices \code{x} and \code{y} for a given length scale parameter \code{l}.
}
