% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfnct00.R
\name{MH}
\alias{MH}
\title{A Metropolis-Hastings}
\usage{
MH(initial, mean, covmat, nu, q, M1, n)
}
\arguments{
\item{initial}{Initial vector of outputs and derivatives.}

\item{mean}{Stacked vector mean of the GP and its derivative.}

\item{covmat}{Stacked covariance matrix.}

\item{nu}{Constraint parameter - controls the monotonicity.}

\item{q}{Scale parameter for the covariance matrix of the proposal distribution.}

\item{N}{Number of MCMC iterations.}
}
\value{
A list: first element is a draw from the constrained GP. Second element is the
acceptance rate of the MCMC.
}
\description{
Metropolis-Hastings step to draw output and derivative from the constrained GP with fixed GP parameters
}
