% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfnct00.R
\name{sampling}
\alias{sampling}
\title{MCMC sampling within SMC}
\usage{
sampling(l, sig2, y, yd, M, nu)
}
\arguments{
\item{l}{Correlation parameter}

\item{sig2}{Variance}

\item{y}{Vector of responses}

\item{yd}{Vector of derivatives}

\item{M}{Number of MCMC transition, default is 1.}

\item{nu}{Constraint parameter}
}
\value{
list of revived samples for each parameter
}
\description{
MCMC sampling within SMC
}
