% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMC.R
\name{SMC_MonGP}
\alias{SMC_MonGP}
\title{SMC sampling to generate monotone GP sample paths}
\usage{
SMC_MonGP(x, y_obs, xd, newx, N = 2000, L = 10, initial_values)
}
\arguments{
\item{x}{Evaluation sites}

\item{y_obs}{Evaluations}

\item{xd}{virtual derivative sites}

\item{newx}{prediction sites}

\item{N}{Number of particles}

\item{L}{Number of densities in the sequence}

\item{initial_values}{a named list of initial values for the covariance parameters \code{l} 
and \code{sig2}}
}
\value{
a named list: samples drawn from the sequence of posteriors for each parameter
}
\description{
SMC sampling to generate monotone GP sample paths
}
\examples{
x <- as.matrix(c(seq(0,.4,.1), seq(.9,1,.1)))
fnct <- function(x) {
  out <- log(20 * x + 1)
  return(out)
}
data <- fnct(x)
mu <- mean(data)
se <- as.numeric(sqrt(var(data)))
y_obs <- (data - mu)/se
newx <- as.matrix(seq(.01,.99,length=50))
xd <- as.matrix(seq(.42,.87,length=10))
initial_values = list(l = .5, sig2 = 10)
obj = SMC_MonGP(x, y_obs, xd, newx, N = 2000, L = 10, initial_values)
sample_path(obj) 
}
